requirements.txt
flask
mysql-connector
gunicorn
flask_restful